Added JNodes_ModelInOut
This node is a simple input-output for a checkpoint model type. This takes in a model and outputs the same model with no processing or changes made. This is intended only to increase compatibility with KJNodes Set/Get nodes.

The set nodes cannot be connected directly to a bypassed node or dynamically typed nodes, so if you try to connect a bypassed group of nodes (like say, switching ipa model processing on and off) to the set node, it won't work. Adding a dynamically types in-out node also won't work for the same reason, so it was necessary for my workflow to add this node for cleanliness. Maybe others will find it useful as well.